Renamed method for user exercise data
