+++ src/main/java/seedu/duke/Duke.java
@@ -7,6 +7,8 @@ import seedu.duke.userdata.UserCareerData;
 import seedu.duke.ui.Ui;
 import seedu.duke.states.ExerciseStateHandler;
 import seedu.duke.userplan.UserPlan;
+
+import java.util.HashMap;
 import java.util.Scanner;
 
 public class Duke {
@@ -16,10 +18,10 @@ public class Duke {
     private final GenerateExercise exerciseGenerator;
     private final ExerciseStateHandler exerciseHandler;
     private UserCareerData userCareerData;
+    private HashMap<String,Integer> userExerciseHistory;
     private final StorageHandler storageHandler;
 
     public Duke () {
-
         ui = new Ui();
         planner = new UserPlan();
         exerciseGenerator = new GenerateExercise();
