+++ src/main/java/seedu/duke/commandhandler/CommandHandler.java
@@ -35,5 +35,4 @@ public class CommandHandler {
     }
 
 
-
 }
