+++ src/main/java/seedu/duke/commandhandler/ExerciseSessionCommandHandler.java
@@ -10,7 +10,9 @@ import seedu.duke.exceptions.OngoingExWriteError;
 import seedu.duke.states.ExerciseStateHandler;
 import seedu.duke.ui.Ui;
 import seedu.duke.userdata.UserCareerData;
+import seedu.duke.userexercisedata.UserExerciseData;
 
+import java.util.HashMap;
 import java.util.Scanner;
 
 public class ExerciseSessionCommandHandler implements CommandList{
@@ -67,6 +69,10 @@ public class ExerciseSessionCommandHandler implements CommandList{
                 break;
             case HISTORY_COMMAND:
                 throw new OngoingExHistoryError();
+            case EXERCISE_DATA_COMMAND:
+                HashMap<String, Integer> userExerciseDataMap = UserExerciseData.addUserExerciseHistory(userCareerData);
+                ui.printUserExerciseHistory(userExerciseDataMap);
+                break;
             default:
                 ui.unknownCommand();
                 break;
